Provide default connection string for testing
This eliminates the need to set a repository secret for the GitHub action, which fails when the action is running as the result of a pull request from a forked repo.

Also updated the start-db.sh script to use pg_ready to wait for postgres to become ready instead of sleeping for a predefined duration.